mdast-util-split: rename to mdast-util-split-by-heading; README review
