+++ README.md
@@ -35,7 +35,7 @@ This project uses [Jest][jest] for testing. It is recommended to install the Jes
 
 ## Packages
 
-* [**mdast-util-split**][mdast-util-split]
+* [**mdast-util-split-by-heading**][mdast-util-split-by-heading]
 
   A MDAST tool to split a markdown tree into list of subtrees representing the chapters. It relies on heading depth.
 
@@ -152,7 +152,7 @@ This project uses [Jest][jest] for testing. It is recommended to install the Jes
 
 [jest]: https://facebook.github.io/jest/
 
-[mdast-util-split]: https://github.com/zestedesavoir/zmarkdown/tree/master/packages/mdast-util-split#mdast-util-split--
+[mdast-util-split-by-heading]: https://github.com/zestedesavoir/zmarkdown/tree/master/packages/mdast-util-split-by-heading#mdast-util-split-by-heading--
 [rebber]: https://github.com/zestedesavoir/zmarkdown/tree/master/packages/rebber#rebber--
 [rebber-plugins]: https://github.com/zestedesavoir/zmarkdown/tree/master/packages/rebber-plugins#rebber-plugins--
 [remark-abbr]: https://github.com/zestedesavoir/zmarkdown/tree/master/packages/remark-abbr#remark-abbr--
