+++ packages/mdast-util-split-by-heading/.babelrc
@@ -1,5 +1,5 @@
 {
-  "presets": ["env"],
+  "presets": ["@babel/preset-env"],
   "plugins": [],
   "env": {
     "development": {
