+++ packages/mdast-util-split-by-heading/.eslintignore
