+++ packages/mdast-util-split-by-heading/.eslintrc.json
