+++ packages/mdast-util-split-by-heading/LICENSE-MIT
