+++ packages/mdast-util-split-by-heading/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "mdast-util-split",
-  "version": "1.0.0",
+  "name": "mdast-util-split-by-heading",
+  "version": "0.0.1",
   "description": "Split MDAST into subtrees relying on the header hierarchy.",
-  "repository": "https://github.com/zestedesavoir/zmarkdown/tree/master/packages/mdast-util-split",
+  "repository": "https://github.com/zestedesavoir/zmarkdown/tree/master/packages/mdast-util-split-by-heading",
   "author": "François (artragis) Dambrine  <perso@francoisdambrine.me>",
   "contributors": [],
   "scripts": {
