+++ packages/mdast-util-split-by-heading/src/index.js
